chore: sets proportions code separated in data generation and graphs

+++ TP3/config.py
@@ -12,12 +12,12 @@ def load_config():
 def ex2_test_size():
     with open('config.json') as json_file:
         config = json.load(json_file)
-        return int(config['ex2_test_size'])
+        return float(config['ex2_test_size'])
 
 def ex3_test_size():
     with open('config.json') as json_file:
         config = json.load(json_file)
-        return int(config['ex3_test_size'])
+        return float(config['ex3_test_size'])
 
 def ex3c_noise():
     with open('config.json') as json_file:
